requirements.txt
streamlit
pdfplumber
python-docx
spacy
sentence-transformers
transformers
openai
python-dateutil
sentencepiece
